talker_service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "talker-service"
version = "0.1.0"
description = "Python service for TALKER Expanded - handles AI compute for NPC dialogue"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "TALKER Expanded Contributors"}
]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "tornado>=6.1",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "neo4j>=5.26.0",
]

[project.optional-dependencies]
stt = [
    "faster-whisper>=1.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "soundfile>=0.12.0",
    "av>=13.0.0",
]
tts = [
    "pocket_tts>=0.1.0",
    "imageio-ffmpeg>=0.4.0",
    "numpy>=1.24.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-timeout>=2.3.0",
    "respx>=0.20.0",
    "mcp>=1.0.0",
    "pytest-json-report>=1.5.0",
    "jsonschema>=4.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
talker-service = "talker_service.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
timeout = 10
timeout_method = "thread"
